requirements.txt
django-autoslug
django-taggit
django-bootstrap-toolkit
django-tables2
django-braces
Pillow
elasticsearch
pyelasticsearch
Pygments
dictdiffer
django-autocomplete-light
django-datetime-widget
winappdbg
django-filebrowser
django-ckeditor
django-tinymce
sorl-thumbnail
Django-tinymce-filebrowser
django-simple-captcha
django-contrib-comments
django-mptt
xlrd